==> R/00_do.R <==
# Run all scripts --------------------------------------------------------------
source(file = "R/01_load_and_clean.R")
source(file = "R/02_augment.R")
source(file = "R/03_analysis_descriptive.R")
source(file = "R/04_analysis_SIR.R")
source(file = "R/05_analysis_symptoms.R")
